Add missing JS libs required for a cmd line build
We would like to add the Berlin-Vegan Guide
app to the F-Droid store. For this purpose
we need to be able to do an automated build
of the app from the sources.

The two JavaScript "libraries" cordova.js and
cordova_plugins.js are required in lib/android if
you want to perform a command line build for
Android (i.e. without using IntelliJ).

cordova.js was taken from

    <https://github.com/apache/cordova-js/blob/master/src/cordova.js>
    (cf2ab5190f94de8923aafd5559be940661a32e75)

and cordova_plugins.js was taken from

    <https://github.com/phonegap/phonegap-2-style-3/blob/master/platforms/android/assets/www/cordova_plugins.json>
    (c372f70a93c7fc4f98bd813c1b00504d4bb59f90)